pyproject.toml
[tool.poetry]
name = "spark-playground"
version = "0.1.0"
description = ""
authors = ["lvittor <lvittor@itba.edu.ar>"]

[tool.poetry.dependencies]
python = "^3.10"
pyspark = "3.1.2"
python-twitter = "^3.5"
apache-airflow = "^2.4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
